pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "golem-agent"
version = "0.1.0"
description = "An autonomous AI agent — picks up work, executes it, validates the results, and reports back."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["ai", "agent", "automation", "claude", "llm", "task-execution"]
authors = [
    { name = "Boris Sobol" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "requests>=2.28",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "pylint>=3.0",
]

[project.urls]
Homepage = "https://github.com/itsmeboris/golem"
Repository = "https://github.com/itsmeboris/golem"
Issues = "https://github.com/itsmeboris/golem/issues"

[project.scripts]
golem = "golem.cli:main"

[tool.setuptools.packages.find]
include = ["golem*"]

[tool.pylint."messages control"]
disable = [
    "import-outside-toplevel",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "duplicate-code",
    "too-many-arguments",
]

[tool.pylint.format]
max-line-length = 99

[tool.pylint.main]
load-plugins = []
source-roots = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["golem/tests"]
addopts = "-m 'not integration'"
markers = [
    "slow: long-running tests",
    "integration: tests requiring external services or API budget",
]

[tool.coverage.run]
source = ["golem"]
omit = ["golem/tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
